LPS-155444 RenameAppPropertiesContext directory
