+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/CardSubscription/index.js
@@ -12,7 +12,7 @@
 import {useModal} from '@clayui/modal';
 import {useState} from 'react';
 import i18n from '../../../../common/I18n';
-import {useAppPropertiesContext} from '../../../../common/context/AppPropertiesProvider';
+import {useAppPropertiesContext} from '../../../../common/contexts/AppPropertiesContext';
 import ModalCardSubscription from '../../containers/ModalCardSubscription';
 import {STATUS_TAG_TYPES} from '../../utils/constants';
 import getDateCustomFormat from '../../utils/getDateCustomFormat';
