+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/DeactivateKeysTable/components/DeactivateButton/index.js
@@ -14,7 +14,7 @@ import {useModal} from '@clayui/modal';
 import {useState} from 'react';
 import {useNavigate} from 'react-router-dom';
 import i18n from '../../../../../../common/I18n';
-import {useAppPropertiesContext} from '../../../../../../common/context/AppPropertiesProvider';
+import {useAppPropertiesContext} from '../../../../../../common/contexts/AppPropertiesContext';
 import {putDeactivateKeys} from '../../../../../../common/services/liferay/rest/raysource/LicenseKeys';
 import {ALERT_DOWNLOAD_TYPE, STATUS_CODE} from '../../../../utils/constants';
 import ConfirmationMessageModal from '../../../ActivationKeysTable/components/Deactivate/ConfirmationMessageModal';
