+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/TeamMembersTable/components/PopoverIconButton/index.js
@@ -13,7 +13,7 @@ import ClayButton from '@clayui/button';
 import ClayIcon from '@clayui/icon';
 import ClayPopover from '@clayui/popover';
 import i18n from '../../../../../../common/I18n';
-import {useAppPropertiesContext} from '../../../../../../common/context/AppPropertiesProvider';
+import {useAppPropertiesContext} from '../../../../../../common/contexts/AppPropertiesContext';
 const PopoverIconButton = ({alignPosition = 'bottom'}) => {
 	const {articleAccountSupportURL} = useAppPropertiesContext();
