+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/onboarding/pages/Welcome/index.js
@@ -12,7 +12,7 @@
 import i18n from '../../../../common/I18n';
 import {Button} from '../../../../common/components';
 import Layout from '../../../../common/containers/setup-forms/Layout';
-import {useAppPropertiesContext} from '../../../../common/context/AppPropertiesProvider';
+import {useAppPropertiesContext} from '../../../../common/contexts/AppPropertiesContext';
 import {useOnboarding} from '../../context';
 import {actionTypes} from '../../context/reducer';
 import {ONBOARDING_STEP_TYPES} from '../../utils/constants';
